requirements.txt
streamlit==1.28.0
opencv-python-headless==4.8.1.78
mediapipe==0.10.7
pillow==10.0.1
numpy==1.24.3
